feat(sdk-coin-trx): fix verifyTransaction method for trx tokens
BG-66457
